refactor: move PhaseTimeline component to it's own file
